Reorganize repository into clear feature-based folders
