+++ Core/ConfigHelper.cs
