+++ Forms/Employee/updateemp.cs
@@ -1,0 +1,71 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Data.SqlClient;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace Employeemanagment
+{
+    public partial class updateemp : Form
+    {
+        SqlConnection Con = new SqlConnection(ConfigHelper.GetConnectionString());
+        public updateemp()
+        {
+            InitializeComponent();
+        }
+
+        private void updateemp_Load(object sender, EventArgs e)
+        {
+
+        }
+
+        private void button1_Click(object sender, EventArgs e)
+        {
+            admin add = new admin();
+            add.Show();
+            this.Hide();
+        }
+
+        private void button2_Click(object sender, EventArgs e)
+        {
+            Form1 f1 = new Form1();
+            f1.Show();
+            this.Hide();
+        }
+
+        private void button3_Click(object sender, EventArgs e)
+        {
+            Application.Exit();
+        }
+        private void populate()
+        {
+            Con.Open();
+            string query = "select * from employee";
+            SqlDataAdapter sda = new SqlDataAdapter(query, Con);
+            SqlCommandBuilder builder = new SqlCommandBuilder(sda);
+            var ds = new DataSet();
+            sda.Fill(ds);
+            allemployeeDGV.DataSource = ds.Tables[0];
+            Con.Close();
+        }
+        private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
+        {
+
+        }
+
+        private void panel1_Paint(object sender, PaintEventArgs e)
+        {
+            populate();
+        }
+
+        private void button4_Click(object sender, EventArgs e)
+        {
+
+        }
+    }
+}
