+++ README.md
@@ -1,5 +1,17 @@
 # Employee-Management-system-
-2nd year Software development project 
+2nd year Software development project.
+
+## Project Structure
+
+- `Core/` - core startup and shared configuration helper classes.
+- `Configuration/` - runtime configuration files (`appsettings.json`).
+- `Forms/Main/` - landing and entry forms.
+- `Forms/Auth/` - login-related forms.
+- `Forms/Admin/` - admin dashboard form.
+- `Forms/Employee/` - employee management forms (add/search/update/delete/salary/shift/info).
+- `Assets/Icons/` - icon files used by the application.
+- `Database/` - SQL scripts.
+- `Docs/` - supporting project documents.
 
 ## License
