--- searchemp.cs
@@ -1,70 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Data.SqlClient;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace Employeemanagment
-{
-    public partial class searchemp : Form
-    {
-        public searchemp()
-        {
-            InitializeComponent();
-        }
-        SqlConnection Con = new SqlConnection(@"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=C:\Users\admin\Documents\employeeDb.mdf;Integrated Security=True;Connect Timeout=30");
-
-        private void button1_Click(object sender, EventArgs e)
-        {
-            admin add = new admin();
-            add.Show();
-            this.Hide();
-        }
-
-        private void button2_Click(object sender, EventArgs e)
-        {
-            Form1 f1 = new Form1();
-            f1.Show();
-            this.Hide();
-        }
-        private void populate()
-        {
-            Con.Open();
-            string query = "select * from employee where ID='" + IDTb.Text + "'";
-            SqlDataAdapter sda = new SqlDataAdapter(query, Con);
-            SqlCommandBuilder builder = new SqlCommandBuilder(sda);
-            var ds = new DataSet();
-            sda.Fill(ds);
-            dataGridView1.DataSource = ds.Tables[0];
-            Con.Close();
-        }
-        private void button3_Click(object sender, EventArgs e)
-        {
-            this.Hide();
-        }
-
-        private void button7_Click(object sender, EventArgs e)
-        {
-            if (IDTb.Text == "")
-            {
-                MessageBox.Show("Missing Information!!");
-            }
-            else
-            {
-                try
-                {
-                    populate();
-                }
-                catch (Exception Ex)
-                {
-                    MessageBox.Show(Ex.Message);
-                }
-            }
-        }
-    }
-}
