--- shiftemp.cs
@@ -1,79 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Data.SqlClient;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace Employeemanagment
-{
-    public partial class shiftemp : Form
-    {
-        public shiftemp()
-        {
-            InitializeComponent();
-        }
-        SqlConnection Con = new SqlConnection(@"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=C:\Users\admin\Documents\employeeDb.mdf;Integrated Security=True;Connect Timeout=30");
-        private void button1_Click(object sender, EventArgs e)
-        {
-            admin add = new admin();
-            add.Show();
-            this.Hide();
-        }
-        private void button2_Click(object sender, EventArgs e)
-        {
-            Form1 f1 = new Form1();
-            f1.Show();
-            this.Hide();
-        }
-
-        private void button3_Click(object sender, EventArgs e)
-        {
-            Application.Exit();
-        }
-
-        private void panel4_Paint(object sender, PaintEventArgs e)
-        {
-
-        }
-
-        private void label14_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void shiftTb_SelectedIndexChanged(object sender, EventArgs e)
-        {
-
-        }
-        private void button4_Click(object sender, EventArgs e)
-        {
-            if (shiftTb.Text == "")
-            {
-                MessageBox.Show("Enter the Shift who are work now!!");
-            }
-            else
-            {
-                try
-                {
-                    Con.Open();
-                    string query = "select name, ID, shift from employee where shift='" + shiftTb.Text + "';";
-                    SqlDataAdapter sda = new SqlDataAdapter(query, Con);
-                    SqlCommandBuilder builder = new SqlCommandBuilder(sda);
-                    var ds = new DataSet();
-                    sda.Fill(ds);
-                    dataGridView1.DataSource = ds.Tables[0];
-                    Con.Close();
-                }
-                catch (Exception Ex)
-                {
-                    MessageBox.Show(Ex.Message);
-                }
-            }
-        }
-    }
-}
